requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idanalyzer==1.2.2
idna==3.4
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.7.0
urllib3==2.0.3
